Implement simple filesystem backend
