+++ test/inputs/test-fs/a/a.txt
@@ -1,0 +1,1 @@
+quux
