+++ test/inputs/test-fs/b.txt
@@ -1,0 +1,1 @@
+foo
