+++ test/inputs/test-fs/c.txt
@@ -1,0 +1,1 @@
+bar
